+++ conda_smithy/feedstock_outputs.py
@@ -5,13 +5,33 @@ import tempfile
 import os
 import json
 import hmac
-from contextlib import redirect_stderr, redirect_stdout
 import urllib.parse
 
 import requests
 
 import git
-import scrypt
+
+from .feedstock_tokens import is_valid_feedstock_token
+
+
+def feedstock_outputs_are_valid(
+    user, project, outputs, feedstock_token, conda_channel, output_repo, token_repo, register=True
+):
+    if not is_valid_feedstock_token(user, project, feedstock_token, token_repo):
+        return False, "invalid feedstock token"
+
+    valid_outputs = is_valid_feedstock_output(
+        user, project, [o["name"] for _, o in outputs.items()], output_repo, register=register)
+
+    valid_hashes = is_valid_output_hash(conda_channel, outputs)
+
+    for o in outputs:
+        if not valid_outputs[outputs[o]["name"]]:
+            return False, "output %s not allowed for %s" % (outputs[o]["name"], project)
+        if not valid_hashes[o]:
+            return False, "output %s does not have a valid md5 checksum" % o
+
+    return True
 
 
 def is_valid_output_hash(conda_channel, outputs):
@@ -115,76 +135,5 @@ def is_valid_feedstock_output(user, project, outputs, output_repo, register=True
         if register and made_commit:
             repo.remote().pull(rebase=True)
             repo.remote().push()
-
-    return valid
-
-
-def is_valid_feedstock_token(user, project, feedstock_token, token_repo):
-    """Test if the input feedstock_token is valid.
-
-    All exceptions are swallowed and stdout/stderr from this function is
-    redirected to `/dev/null`. Sanitized error messages are
-    displayed at the end.
-
-    If you need to debug this function, define `DEBUG_FEEDSTOCK_TOKENS` in
-    your environment before calling this function.
-    """
-    from .github import gh_token
-
-    github_token = gh_token()
-
-    failed = False
-    err_msg = None
-    valid = False
-
-    # capture stdout, stderr and suppress all exceptions so we don't
-    # spill tokens
-    with tempfile.TemporaryDirectory() as tmpdir, open(
-        os.devnull, "w"
-    ) as fp, redirect_stdout(fp), redirect_stderr(fp):
-        try:
-            # clone the repo
-            _token_repo = (
-                token_repo.replace("$GITHUB_TOKEN", github_token)
-                .replace("${GITHUB_TOKEN}", github_token)
-                .replace("$GH_TOKEN", github_token)
-                .replace("${GH_TOKEN}", github_token)
-            )
-            git.Repo.clone_from(_token_repo, tmpdir, depth=1)
-            token_file = os.path.join(
-                tmpdir, "tokens", project.replace("-feedstock", "") + ".json",
-            )
-
-            # don't overwrite existing tokens
-            # check again since there might be a race condition
-            if os.path.exists(token_file):
-                with open(token_file, "r") as fp:
-                    token_data = json.load(fp)
-
-                salted_token = scrypt.hash(
-                    feedstock_token,
-                    bytes.fromhex(token_data["salt"]),
-                    buflen=256).hex()
-
-                valid = hmac.compare_digest(
-                    salted_token,
-                    bytes.fromhex(token_data["hashed_token"]),
-                )
-        except Exception as e:
-            if "DEBUG_FEEDSTOCK_TOKENS" in os.environ:
-                raise e
-            failed = True
-    if failed:
-        if err_msg:
-            raise RuntimeError(err_msg)
-        else:
-            raise RuntimeError(
-                (
-                    "Registering the feedstock token for %s/%s failed!"
-                    " Try the command locally with DEBUG_FEEDSTOCK_TOKENS"
-                    " defined in the environment to investigate!"
-                )
-                % (user, project)
-            )
 
     return valid
